Refactor into audio, FFT, and render modules
